move imshowhs functions into their own module
